Neighbouring Cell Info, Ciphering Indicator, Samsung MultiRil Implementation
Full Credits to Alexey Illarionov for his implementation of the Replicant libsecril-client-sap
methods to enable access to the OEM_HOOK_RAW methods through the MultiClient interface, please
visit Alexey's Github here: https://github.com/illarionov/SamsungRilMulticlient and star his
work without him this would not be possible.

Advanced Cell Fragment added to display the Neighbouring Cell information in two ways firstly
through telephony manager methods which does not work on Samsung Devices, a fallback is available
through the methods developed by Alexey and will display if these are successful.

Ciphering Indicator also uses Alexey's methods and will display on Samsung devices.

Further manipulation of the results provided will enable an increase in the functionality of
attempting to detect IMSI activity and will be developed over time.

Slight change to the preference to regarding the persistent service operation changing the
preference to ENABLE persistent operation which is defaulted to ON.