AT Command Injection Started, Code Cleanup/Updates, Java 1.7 Language Updates
Extension of the Samsung MultiRil to attempt hooking of the OemHookStrings method,
presently I am not 100% certain this will work but the basic framework and extensions
have been made to modify the current implementation to deal with the Strings request
data and responses.

This still requires work and as such is currently disabled until further testing
has been completed as it will cause the socket to disconnect.

Huge code cleanup and updates to a number of areas with optimisations across a large
number of packages, some unused imports removed and code comments added to a variety
of methods.

Addressed a number of possible NPE causes through different parts of the application
which will ultimately provide better stability and an increased user experience,
although in saying that these items were not causing any issues that had been reported
but prevention is better then a cure right?!?! :)

Use of Java 1.7 language rules have further allowed updates to take advantage of
the highler langauge level with removal of explicit types within arrays, lists, maps
etc plus other changes.

Unnecessary compatibility library removed which has resulted in a reduction in size
of the APK.

Reversed the previous change to insert the Cell ID and Lac in Hex format to allow
further time to ensure all methods that utilise these variables to be updated, the
original change was made too quickly and resulted in a number of issues so I have
returned to the standard integer format until I have more time to ensure this
change is implemented correctly.